refactor: move 'containers' models into publishing
