+++ openedx_learning/api/authoring.py
@@ -11,7 +11,6 @@ APIs.
 # pylint: disable=wildcard-import
 from ..apps.authoring.collections.api import *
 from ..apps.authoring.components.api import *
-from ..apps.authoring.containers.api import *
 from ..apps.authoring.contents.api import *
 from ..apps.authoring.publishing.api import *
 from ..apps.authoring.units.api import *
