+++ openedx_learning/api/authoring_models.py
@@ -9,8 +9,7 @@ consistent.
 # pylint: disable=wildcard-import
 from ..apps.authoring.collections.models import *
 from ..apps.authoring.components.models import *
-from ..apps.authoring.containers.models import *
 from ..apps.authoring.contents.models import *
-from ..apps.authoring.publishing.model_mixins import *
+from ..apps.authoring.publishing.model_mixins.publishable_entity import *
 from ..apps.authoring.publishing.models import *
 from ..apps.authoring.units.models import *
