+++ openedx_learning/apps/authoring/publishing/apps.py
@@ -14,3 +14,12 @@ class PublishingConfig(AppConfig):
     verbose_name = "Learning Core > Authoring > Publishing"
     default_auto_field = "django.db.models.BigAutoField"
     label = "oel_publishing"
+
+    def ready(self):
+        """
+        Register Container and ContainerVersion.
+        """
+        from .api import register_content_models  # pylint: disable=import-outside-toplevel
+        from .models import Container, ContainerVersion  # pylint: disable=import-outside-toplevel
+
+        register_content_models(Container, ContainerVersion)
