+++ projects/dev.py
@@ -35,7 +35,6 @@ INSTALLED_APPS = (
     "openedx_learning.apps.authoring.components.apps.ComponentsConfig",
     "openedx_learning.apps.authoring.contents.apps.ContentsConfig",
     "openedx_learning.apps.authoring.publishing.apps.PublishingConfig",
-    "openedx_learning.apps.authoring.containers.apps.ContainersConfig",
     "openedx_learning.apps.authoring.units.apps.UnitsConfig",
     # Learning Contrib Apps
     "openedx_learning.contrib.media_server.apps.MediaServerConfig",
